migrate unit_states_test_network to pytest
